Move mixpanel event function to analytics file
